wip: First draft implementation of segmented VoiceHD design
This commit is capable of creating an VoiceHD accelerator that works on
segmented HVs. Some combinations of HV_DIM and HV_SEGMENTS are
HLS-synthesizable. The problem now is how to prevent the HLS tool to
apply transformations that result in a variable with more than 4096
signals. Besides that, the synthesis time was drastically reduced and it
was possible to cosimulate a design with HV_DIM=1000 and
HV_SEGMENT_SIZE=5.