wip: Draft segmented VoiceHD with low latency
This commit enhances voicehd.tcl with directives to optimize the design.
It was possible to drastically reduce latency. However, IM partition is
the bottleneck and I believe that using std::array to implement the IM
may not be a good thing since Vitis seems to have a hard time working
with it. Next commits should refactor C++ arrays to plain C arrays.