New parallel datapath implementation
Now it is possible to scale multiple datapaths. The idea is to create a
bank of accumulators, each one used privately by a datapath. Each
datapath must be instantiated optimized to its bank. To accomplish this,
it is necessary to introduce the "function_instantiate" directive and a
"datapath_id" parameter.