Add bnb operation
Add a new implementation of "bind and bunle" operations. This new
operator performs a fused bind-and-bundle by accumulating bound vectors.
The user must use bsc_bnb_threshold() to convert the accumulated vector
back to its binary form.

This commit also introduces two new data types: bnb_acc_t and
bnb_acc_elem_t. The first is a bank of accumulators of element type
bnb_acc_elem_t, while the second is the data type of each individual
element. Some improvement can be made in bnb_acc_elem_t as it is not
possible for the user to set the precision bits of the accumulators.